+++ assets/templates/wizards/designer/room.c
@@ -9,7 +9,7 @@
  * @doc /doc/build/room/Basic
  * @doc /doc/build/room/Intermediate
  * @doc /doc/build/room/Advanced
-{doc} */
+ {doc} */
 #include <std.h>
 #include "area.h"{includes}
 
@@ -18,6 +18,4 @@ inherit {inherit};{inherits}
 {create pre}void create()
 {
    ::create({create arguments});
-   set_short({short});
-   set_long({long});
 {create body}}{create post}