+++ changelog.md
@@ -25,6 +25,7 @@
       - Cancel on error when generating code
       - Remove buckler as an accessory for armors
       - Fixed set_long string wrapping
+      - Fixed room set_long/short when empty
   - Progress dialog: fix task bar progress not being cleared on close
   - Status: Attempt to fix weather overlay again
   - Display: Set line-height to better control unicode